-don't generate any files in the source tree, but only in the build tree
Laurent, where did this cause problems ?
Was this with the files where source files with the same base names are now
compiled to the same target, and beforfe with libtool one of them was part
of a convenience libs ?

Now that we don't use convenience libs anymore, this means that if two
source files have the same basename, they will generate moc files with the
same name, also for ui-files. What can we do about this ?
If two files have the same names, maybe the name is too generic, so let's
rename one of them to a more specialized name, especially now that we have
svn.

But generating files is the source-tree is a no-no.

CCMAIL: montel@kde.org


svn path=/trunk/KDE/kdelibs/; revision=520047