try to fix the automoc problem in kioslave/media/medianotifier
kde4_automoc() absolutely _must_ work without the macro_add_file_dependencies() line.
If it doesn't it's a cmake bug. But since this is the way the dependencies of generated
files are handled everywhere I am at least 99% sure it works in cmake and the mistake must be on my side.
So, I think the problem was that mediamanageriface.cpp was also used from mdeia/medianotifier/ using "../mediamanageriface.cpp".
Then this file was parsed for "include *moc" and probably the correct line was found and added as a property
to the source file using the absolute path. Later on in kde4_add_plugin() the property for the moc
files was retrieved but without using the absolute path, so that's why probably the moc file 
wasn't found and so the dependency wasn't created.

CCMAIL: faure@kde.org

David, can you please check and let me know ?

Thanks
Alex



svn path=/trunk/KDE/kdelibs/; revision=598892