Undo my accidental change of the compiler flags from the other day. So, re-add the KDE 3.5.x test to check if g++ is fixed WRT to -fvisibility-inlines-hidden and enable it if it works (works fine on 32-bit).
As for the C compiler flag, I've moved it to an independent
variable. KDE is not in the business of creating C libraries, so I'll
add this flag for kde4_add_library & family. If we need a 3rd-party
library to be build, a different macro than kde4_add_library will be
needed.

svn path=/trunk/KDE/kdelibs/; revision=626277