It is really a bad idea to search for stuff _only_ using NO_DEFAULT_PATH. In the case that what you are looking for is somewhere else then cmake has no way to find it. For custom install locations cmake provides the environment variables CMAKE_LIBRARY_PATH and FIND_INCLUDE_PATH. So if you have your own set of install locations, set these two environment variables accordingly. cmake cvs additionally offers CMAKE_FIND_PREFIX_PATH, which is used by all FIND_PATH/FILE/LIBRARY/PROGRAM() calls with the appropriate subdir appended. I.e. if you install everything to ~/mystuff/, do export CMAKE_FIND_PREFIX_PATH=$HOME/mystuff, cmake will append the include/, lib/ and bin/ subdirectories automatically.
Adding CMAKE_INSTALL_DIR/<suffix> is not necessary, this is done by default in FindKDE4Internal.cmake
(and in cmake > 2.4.7).

So now at first the FIND_XXX() calls are used with NO_DEFAULT_PATH, so no change here for UNIX, and after
that the FIND_XXX() calls are used without NO_DEFAULT_PATH for the case that the first one didn't succeed (so no change for Windows).
This has the nice effect that the cmake-provided means to specify custom search locations work again.


Alex

CCMAIL: ch.ehrlicher@gmx.de
CCMAIL: kretz@kde.org


svn path=/trunk/KDE/kdelibs/; revision=741164