backport SVN commit 764809 by winterz:
patch by dfaure so that the testfoo.shell wrapper is used  when running
'make test' on Unix.  We want to do this because testfoo.shell
uses the runpath that points to our builddir, rather than installdir.

this way, you no longer need to run make install before running make test.

as discussed on kde-buildsystem


svn path=/branches/KDE/4.0/kdelibs/; revision=764816