Fix FindPackageHandleStandardArgs.cmake for cmake's config-mode.
When using the Config-Mode with a -version-file then cmake automatically sets
the <module>_VERSION variable, but it uses the mixed-case that is used as
module name.

The logic in this file so far however always expected the version variable to be
all upper-case breaking the logic when using it in a simplified FindXxx.cmake module.

This happened to kdevplatform/kdevelop.

svn path=/trunk/KDE/kdelibs/; revision=1123271