Check gcc's include path and add any previously missing entries to CMAKE_{C,CXX}_IMPLICIT_INCLUDE_DIRECTORIES.
As discussed on the kde-buildsystem mailing list, this is particularly
useful on systems such as FreeBSD, which have most of the libraries and
headers installed in /usr/local: CMake would not add -I/usr/include when
compiling code, but would do that with other directories such as
/usr/local.

This, in turn, would sometimes make the system-wide includes be added to
the include path before the local ones, which can lead to compilation
failures.

With this patch, if CPATH is properly set, directories such as /usr/local/*
will not be added by CMake when gcc is invoked.

For now, this works only with gcc, but can be extended to other compilers
if necessary.

CCMAIL: kde-buildsystem@kde.org

svn path=/trunk/KDE/kdelibs/; revision=1135642