-add a directory for KDE specific files and add a KDEInstallDirs.cmake
KDEInstallDirs.cmake is similar to GNUInstallDirs.cmake coming with cmake,
but provides the install variables as used by KDE.
It also provides the special feature of initializing them to the values from kdelibs (?)
when installed into the same prefix.

Currently it all still uses "kde4", I'm not sure this will stay this way.

Alex