Rename file and macro for consistency, as requested by Alexander.
"Can you please rename it to ECMDBusAddActivationService.cmake and
the contained macro to ecm_dbus_add_activation_service(), so it follows the
naming conventions in cmake (macro has the same name as the file) and ecm (all
macros have the prefix ecm)." -> Done.
Nothing to adapt in kdelibs-frameworks yet, it has its own copy for now
(I'll get rid of it at the next ECM version upgrade).