Add Clang support to KDECompilerSettings.cmake
This more or less copies what was done for KDE4 in
 https://git.reviewboard.kde.org/r/111612/.

With the changes applied I successfully installed a hello world
application. The linker and compiler command lines seem to be correct
and include all extended features. I also checked and debugfull also
works as expected (-g3 is added).

REVIEW: 111661