Remove commented-out manifest tool code from KDECompilerSettings
This originally found the manifest tool from kdewin, which was then used
by kde4_add_executable to embed a standard manifest file in
applications, apparently allowing applications to have administrator
privileges when run by an administrator.

Given we do not have kde4_add_executable any more, this is useless (and
certainly does not belong in this file).

kdewin should provide any relevant manifest macros itself, in a
KDEWinMacros.cmake file or some such, and these should be used on an
as-needed basis for executables that require it.