Improve the ECMGenerateHeaders API with a variable for generated files
ecm_generate_headers() now allows/forces the caller to collect the paths
of the generated headers, so that they can be passed to the install
command.  This avoids issues of unexpected files being in the CamelCase
includes directory, both from previous builds and because of
case-insensitive file systems.

MODULE_NAME is removed, as it is no longer desirable or necessary.
Instead, the headers are placed directly in the output directory
(usually CMAKE_CURRENT_BUILD_DIR).

Overall, this makes ecm_generate_headers() behave much more like other
file generation macros (like the Qt ones).

The old syntax is still supported for now, to make the porting effort
easier.

REVIEW: 115765