Remove ECMDBusAddActivationService module
ecm_dbus_add_activation_service() requires suffient knowledge of its
internals to use that replacing two lines with one seems silly.

In order to use it you have to know it behaves like configure_file()
(because you have to construct the file yourself), except that it also
installs it somewhere (for which you have to make sure
DBUS_SERVICES_INSTALL_DIR is defined before you use it, which is
certainly not a given for non-KDE projects). By this point, why not just
use configure_file() and install()? The DBUS_SERVICES_INSTALL_DIR
provided by KDEInstallDirs is all the magic you actually need, and if
that's explicit in the CMakeLists.txt file, it's a lot more obvious that
you should have it defined somewhere.

REVIEW: 117581