New syntax for ecm_install_icons()
This requires the icon files to be specified (which is better than
globbing, because the build system will then be able to tell when files
are added or removed and re-run CMake).

It also removes the theme name from the filename pattern: the old code
used a shorthand theme name for a small number of themes, and didn't
allow any other themes. Extending this to arbitrary themes could cause
problems with themes that have numbers or hyphens (or whatever other
delimiter character was used) in their names. Most users are likely to
just want to install to a single theme anyway (based on a random
sampling of users of kde4_install_icons), so that is what the new syntax
requires.

The old syntax still works and behaves as before.

ecm_update_iconcache is renamed to _ecm_update_iconcache - it was never
documented as public API anyway.

REVIEW: 117617