Add a stub for ECMFindModuleHelpers to the find-modules dir
ECMUseFindModules allows find modules to be copied to a local directory.
These find modules may use ECMFindModuleHelpers, but they will not be in
the same relative location to ECMFindModuleHelpers.cmake and there is no
guarantee that ECMFindModulesHelpers.cmake will be in the CMake module
path.

To solve this, we make sure there is always a stub file in the same
directory as the find modules that includes the real
ECMFindModuleHelpers.cmake. The one installed with ECM just includes
"../modules/ECMFindModuleHelpers.cmake", while ecm_use_find_modules
generates a stub that uses an absolute path.

REVIEW 117658