Fix finding of Qt5LinguistTools on some systems
Although ECM does not make use of a compiler directly, the language
affects the search path for CMake packages; in particular, a package
installed to /usr/lib64/cmake will not be found if NONE is passed as the
language argument to project(). This meant that a 64-bit version of
Qt5LinguistTools would not be found on systems where 64-bit libraries
are not installed in the "default architecture" location (/usr/lib).

With this change, the configure step performs some otherwise-unnecessary
tests. We minimise this by explicitly specifying the C language (which
is also what some of the tests use), rather than letting it be the
default (which is C and C++).

REVIEW: 118374