Remove Qt5 debugfull hack from ECMConfig.cmake
The original purpose of this was to set the QT_NO_DEBUG macro if the old
DebugFull configuration was used. We got rid of DebugFull (instead just
using Debug), so it is no longer relevant, and this hack never belonged
in ECMConfig.cmake anyway (it should have been in KDECompilerSettings).

CHANGELOG: ECM now works when the required CMake version is set to
3.0.0 or higher, and does not require Qt5Core to be available.

BUG: 331849

REVIEW: 119588