KDECompilerSettings: Be more portable across libc's.
Defining _XOPEN_SOURCE to 500 is too restrictive: it corresponds to
_POSIX_C_SOURCE 199506L, and hides several symbols that standard
libraries like libc++ expect to find, leading to errors like this on
FreeBSD:

  In file included from /tmp/attica/src/accountbalance.cpp:21:
  In file included from /tmp/attica/src/accountbalance.h:25:
  In file included from /usr/local/include/qt5/QtCore/QString:1:
  In file included from /usr/local/include/qt5/QtCore/qstring.h:50:
  In file included from /usr/include/c++/v1/string:437:
  In file included from /usr/include/c++/v1/string:437:
  /usr/include/c++/v1/cstdio:143:9: error: no member named 'snprintf' in the global namespace
  using ::snprintf;
        ~~^

This isn't a problem on Linux (actually, on systems using glibc) because
defining _GNU_SOURCE enables a lot more features that are not made
available on other libc implementations where it does not have any
effect.

Instead, stop defining _XOPEN_SOURCE at all and leave it up to the
platform to show or hide as many symbols as necessary if no
standards-related defines are set, and only set _GNU_SOURCE on systems
where it is actually meaningful (ie. systems using glibc).

REVIEW: 119696