ECMQtDeclareLoggingCategory: Include <QDebug> with the generated file
In order to use the logging categories created by ECM,
the generated header file has to be included but then
also #include <QDebug> must be added to make it usable
at all for qCDebug/qCWarning stuff.

So this puts the #include <QDebug> into the generated
header file and thus allowing to include a single header
for qCDebug/qCWarning instead of two.

REVIEW: 126368