Fix _ecm_update_iconcache to only update the install location
Use the CMAKE_INSTALL_PREFIX to ensure that only a file in the install
location will be touched.  Otherwise a user running

 make install

as root would create root-owned files in the build directory.

Use the cmake -E touch command to update the mtime of the directory.
That command exists since late 2007, but this manual file creation
workaround pre-dates that by a few months (kdelibs SVN revision 691973).

There is no chance any calling code is already passing an absolute path
here because later in the function it is already used together with a
CMAKE_INSTALL_PREFIX.

Inspired-by: REVIEW 127445