Add the rules_engine and sip_generator
These files process C++ headers with libclang and python-clang bindings,
using the AST to generate SIP files for PyQt binding library generation.

The design allows for further extensions which can process other
artifacts found in headers.  This iteration of the design is a simple
minimum which suffices to generate headers for KItemModels and some
other libraries.