Android deployment: support projects without things in share or lib/qml
Just setting the field android-extra-plugins to an empty string resulted
in androiddeployqt complaining about
"External resource  does not exist or not a correct directory!"
so the field is completely left out now if no plugin data dirs are found.
For consistency the same is done with the android-extra-libs field.

REVIEW: 127700
BUG: 362578

Thanks apol for review