The default level for logging categories should be Info rather than Warning.
The whole point of Info is to be used by apps who want to print out some
info for the user, not disabled by default.

Example use case: ksmserver could output the name of the autostart
files it's starting, so user can check their own scripts are started,
and associate any errors with the corresponding script.

In other words:
Debug = debugging for the developer
Info = debugging for the user :-)

This commit shows the benefit of having a central place for changing this
(many category definitions do not use this macro though)

REVIEW: 128232