Make sure ECMGeneratePriFile.cmake behaves like the rest of ECM
In KDEInstallDirs we have some code to make sure that qmake is asked when
the project shares the prefix with the installed Qt, to make sure that if
something was changed in the distribution it would reflect on the projects.

Make sure PRI files are installed using the same reasoning.

REVIEW: 128427