Find the correct path to the cmake command
Summary: When running tests with hardcoded `cmake`, it get's suffixed to the project's build path. Setting it with an absolute path ensures using the right cmake binary.

Test Plan: Compiled, installed. Used to generate tests for kdev-embedded. `cmake` binary is now prefixed with the absolute system path, and no longer with the project's build dir.

Reviewers: kfunk, #frameworks, apol

Reviewed By: kfunk, apol

Subscribers: kfunk

Differential Revision: https://phabricator.kde.org/D3568