Explicitly pass -std=gnu++14 to clang when generating bindings
The generator expression here looks like it should work, but it does not
set the appropriate flags for compilation.  It seems that the
CXX_STANDARD property is not yet populated at the time the generator
expression is evaluated (to be investigated later).

This came to light because users of Qt 5.7+ attempted to generate the
bindings, but they encountered errors.  The step of using libclang to
parse the provided headers was actually failing, but that was not
noticed, perhaps because the logging infrastructure in sip_generator
does not emit it (to investigate later).

BUG: 374801