Bindings: Fix logging output severity for parsing messages
The Python logging module uses logging severities with numerical values
which form a sequence in steps of 10.  The Clang cindex.Diagnostic
numerical values use a step size of 1, so the two are incompatible.

Introduce a mapping function so that appropriate errors get issued when
attempting to build the bindings.  The logging module has one surplus
diagnostic level, but we simply never use it.

BUG: 374801