Bindings: Fix handling of forward declarations
It is not appropriate to decorate each forward declaration with the SIP
attribute /External/.  That is only needed for forward declarations of
types which are defined in a different module.

Local forward declarations can be omitted from the sip code.  In sip
code, a forward declaration followed later by a full class definition is
an error.

Omit forward declarations unless they are decorated with the external
attribute.  Introduce a rules database for consumers to decorate them
with the attribute as required.