Introduce fetch-translations build command
Summary:
Makes it possible to fetch translations when building the project. To do so
it fetches kde:releaseme and uses the new fetchpo.rb program to download the
translations into the build directory.

This should make it much easier to integrate translations in the development
process.

Test Plan: Downloaded and installed translations for some projects

Reviewers: #frameworks, #build_system, kfunk, ltoscano, aacid, ilic, sitter

Reviewed By: sitter

Subscribers: sitter

Tags: #frameworks, #build_system

Differential Revision: https://phabricator.kde.org/D5143