Make ECMPoQmToolsTest actually fail if a translation is wrong
Currently the test still passed even if one of the check_translations()
calls would fail.
The reason is that check_translations() is a function which has its own
scope, so if it sets "failed" the upper level doesn't see it.

Use the PARENT_SCOPE option when setting the variable to properly
propagate the value to the upper level.

Differential Revision: https://phabricator.kde.org/D6701