Include a module for finding qml imports as runtime dependencies
Summary:
Allows to check if a module is available on the system and sets it as a
runtime dependency.
This is useful for projects so that they can specify their qml dependencies
easily and packagers and developers get to see what's missing by looking
at the cmake output.

Reviewers: #build_system, #frameworks, sitter

Reviewed By: sitter

Subscribers: dfaure, aacid

Tags: #frameworks, #build_system

Differential Revision: https://phabricator.kde.org/D7094