Make it possible to build several apk out of a project
Summary:
Refactor the apk-generating code into a separate function, in views of
eventually even make it a module.
It also changes so that if no APK dir is specified, a generic dummy one is
used. Useful for proofs of concept.

Test Plan: Built kate, got kate and kwrite apks

Reviewers: #frameworks, #build_system, vkrause

Reviewed By: vkrause

Subscribers: vkrause

Tags: #frameworks, #build_system

Differential Revision: https://phabricator.kde.org/D12150