Add _XOPEN_SOURCE to C definitions
When building on non-glibc Unix platforms, such as Solaris, NetBSD, and Linux/musl or Linux/uclibc, multiple components of KF5 fail to build. This is because -std=iso9899:1990 is specified (for strict C90 compliance) but _XOPEN_SOURCE is not defined, so none of the POSIX interfaces are exported. I have seen this reported in at least:

* kinit
* kscreenlocker
* plasma-workspace

The attached patch resolves this issue on all our build boxes; additionally, applying it to my glibc builder did not change the already working result.

BUG: 373175

Differential Revision: https://phabricator.kde.org/D8256