Disable alteration of QT_PLUGIN_PATH by ECM when running tests.
On Windows at least this is simply broken and leads to tests failing because plugins from other modules cannot be found at all.
Based on this it is likely that it is broken on other platforms as well, it's just being hidden there by other logic.

CCMAIL: kde-frameworks-devel@kde.org