disable autouic again - it breaks builds
Summary:
autouic (like automoc) assumes that every ui_*.h include statement it finds
relates to a .ui file that needs generating. this is not always true.

e.g. we have software which generates ui_debug.h which is simply
a qloggingcategory header for the UI category of that software which would
then trip up autouic because it would assume there's a .ui file when there
really isn't one.

unfortunately the ui_ assumption cannot be selectively disabled, so we
can't have explicit listing of .ui in source lists

```
set(foo_SRCS foo.cpp foo.ui)
```

without also getting the not particularly compatible include assumptions.
this should be revisited for kf6 since there isn't a technical need for
files to be called ui_*, they could just as well be *_ui or anything else
so as to not clash with autouic assumption.

autorcc does not suffer from this problem so we can leave it enabled, for
now anyway.

Test Plan: sonnet builds again

Reviewers: apol, nicolasfella, cullmann

Reviewed By: apol

Subscribers: kde-frameworks-devel, kde-buildsystem

Tags: #frameworks, #build_system

Differential Revision: https://phabricator.kde.org/D23262