new module ECMSourceVersionControl
Summary:
simply sets a variable when the source is under version control. use it to
auto-enable Debug builds. there are also plans to switch special assertion
logic on in KIO when used from git, so there definitely is a more generic
use case of wanting to control behavior based on whether it the source is
likely used to make a development or production build.

conceivably the module could be used in the future to get git rev-parse or
the like, hence the generic name.

Test Plan: with .git the var is true, without it is false

Reviewers: kde-buildsystem, dfaure

Reviewed By: dfaure

Subscribers: apol, kossebau, kde-frameworks-devel

Tags: #frameworks, #build_system

Differential Revision: https://phabricator.kde.org/D24159