ECMQtDeclareLoggingCategory: gently deal with no categories exported
With some build configurations no catgories might be registered for
a given export id.
Instead of failing hard and thus forcing to catch this situation
explicitly on the caller side, be grateful on the callee side
and just generate an empty file, so the installed file set is consistent.

GIT_SILENT