[android] Allow specifying APK install location
Summary:
This allows `make create-apk` to directly write the APK to /output instead of the cp-with-prefix step in /opt/helpers/create-apk. It's also useful for manual development builds where one would need to copy it to some output location manually or for CI setups that expect the output in a certain location.

If ANDROID_APK_INSTALL_DIR is not set the current behaviour is kept.

Reviewers: #frameworks, #android, apol, vkrause

Reviewed By: #android, apol

Subscribers: kde-frameworks-devel, kde-buildsystem

Tags: #frameworks, #build_system

Differential Revision: https://phabricator.kde.org/D29631