Add a way to pass additional arguments to androiddeployqt
We need this to pass the --release option for binary factory release
builds.

An alternative approach would be to set that here automatically depending
on the CMake build type. Since that however ends up producing unsigned and
thus not immediately usable APKs (or would need a whole set of extra
arguments for proper signing) I'm not sure whether that's most convenient
for local use.