Attempt to find the Play store icon if it's not explicitly set
This relies on the apparently predominant naming pattern for those files,
those of our apps not using this often do not seem to have an appropriate
file (512x512 png) to begin with.

Explicit override is also possible, by the already existing mechanism of
putting a file with the right name into the fastlane/ source directory.

Having the icon included here will allow it to be automatically synced to
the Play store.