Provide fallback error when no .clang-format file is present
This check fixes a probably rare case, where we land in a repository
that contains this clang-format.sh hook but where the format file went
missing. In such cases, clang-format falls back to the system defaults,
which brings undesired defaults, e.g. enforced line breaks in comments
after 90 chars, which cannot easily be reverted by a reformat.

Since clang-format's --fallback-style=none option is not available for
git clang-format, this comment reimplements the logic.