Introduce ECMQmlModule.cmake
This contains some helper functions to make it easier to create QML
modules through CMake. It takes care of several things that currently
need to be done manually.

It adds four tests for the four primary ways that it can be used,
either as shared/static library and with or without C++ plugin.