Revert "Fix QT_PLUGIN_PATH for unittests"
As discussed in [1], it's simpler to put the files in builddir/bin/, i.e.
no /plugins subdir, this makes it easier to run the unittest directly from
the CLI/debugger/IDE, without relying on ctest exporting the correct
QT_PLUGIN_PATH; instead QCoreApplication::libraryPaths() will pick up the
plugins in e.g. builddir/bin/kf5/kio/ automatically.

This will be followed by a change in KCoreAddons to remove the /plugins/ dir
from the path in the builddir.

[1] https://invent.kde.org/frameworks/extra-cmake-modules/-/merge_requests/213