Revert "Restore the Android Threads workaround for Qt
We don't need that in regular C/CXX projects, as we break nested CMake
calls with it as e.g. happens during try_compile(). We how do still need
it for CXX-only projects. Fortunately we can check for the language that
early on at least for the cases where this actually matters (the nested
calls).

This should now hopefully fix both kcoreaddons configure check and
kquickimageeditor as a CXX-only project.

This partially reverts commit a85638231ba33653523e3d944912b8f753d94a72.