ECMFindModuleHelpers.cmake:  Generate imported targets with pkg_config and add them to the dependency list.
This allows us to properly link to a static build of XCB.
Parts of XCB depend on libXau, so we must link to libXau, and the
best way to do that is to just use the linker flags returned
by pkg-config.

The minimum required version of CMake to use this feature is 3.6.