Find wayland.xml from wayland-scanner.pc
In wayland, it's the "scanner" meson option that controls whether
wayland.xml is installed, not the "libraries" option (which enables
wayland-client and wayland-server).  This means that, if
wayland-scanner is installed into one path, and wayland-client and
wayland-server are installed into another (as we'd like to do in
Nixpkgs), it's wayland-scanner's pkgdatadir that will contain
wayland.xml.

(cherry picked from commit d7647351c7a10f260643a1b22f9c226437e93aba)