KDE_INSTALL_TARGETS_DEFAULT_ARGS (KF6): drop KDE_INSTALL_INCLUDEDIR
The "INCLUDES DESTINATION dir" argument to "install(TARGETS)" specifies
dirs which will be added to the INTERFACE_INCLUDE_DIRECTORIES property
of the target.

Though KDE_INSTALL_INCLUDEDIR itself should not be a root dir for any
include statements to work, as this dir is shared typically with many
other libraries, so all their headers would be visibile at the same time.
Libraries install their headers explicitely into some visibility-layer
subdir and promote only that as root include dir.
So KDE_INSTALL_INCLUDEDIR added to INTERFACE_INCLUDE_DIRECTORIES is of
no real use.

Keeping around for KDEInstallDirs5 in case some setups rely on it.