ECMGenerateExportHeader: avoid helper C++ macro shared across gen. headers
The macro by the hardcoded name gets redefined in every export header, so
in theory one day could result in definition clashes. As it is just a
helper macro with also unneeded code for the patch level, simply
inlining the logic avoids the need for the macro & results in shorter code,
which still can be grasped by a human reader.